backends/makrx-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
pillow==10.1.0
requests==2.31.0
numpy==1.26.2
trimesh==4.0.5
pycollada==0.8
websockets==12.0
aiofiles==23.2.1
python-magic==0.4.27
pydantic-core==2.14.6
httpx==0.26.0
prometheus-client==0.19.0
structlog==23.2.0

# Feature Flags System Dependencies
# (All already included above, no additional packages needed)